Validation added on categeories deletion - 'Category edition' option added to header

+++ delete_category.php
@@ -15,24 +15,37 @@
 			if (!mysqli_num_rows($result) > 0){
 				$_SESSION['message_error'] = "La categoría no se encontró.";
 				header("Location: categories_management.php");
-			}else{
+			}
+			else {
         $category = mysqli_fetch_array($result);
-				$query = "DELETE FROM `categorias_productos` WHERE `idCategoriaProducto` = '$category_id'";
+
+				// Consultar si existem productos asociados a la categoría
+				$query = "SELECT * FROM `productos` WHERE `idCategoriaProducto` = '$category_id'";
 				$result = mysqli_query($link, $query);
-				mysqli_close($link);
-				if ($result){
-					$_SESSION['message_success'] = "Se eliminó la categoría '" . $category['nombre'] . "'.";
-					header("Location: categories_management.php");
-				}else {
-					$_SESSION['message_error'] = "Error - No se pudo eliminar.";
+				if(mysqli_num_rows($result) > 0){
+					$_SESSION['message_error'] = "Error - No se pudo eliminar porque existen productos asociados.";
 					header("Location: categories_management.php");
 				}
+				else {
+					$query = "DELETE FROM `categorias_productos` WHERE `idCategoriaProducto` = '$category_id'";
+					$result = mysqli_query($link, $query);
+					mysqli_close($link);
+					if ($result){
+						$_SESSION['message_success'] = "Se eliminó la categoría '" . $category['nombre'] . "'.";
+						header("Location: categories_management.php");
+					}else {
+						$_SESSION['message_error'] = "Error - No se pudo eliminar.";
+						header("Location: categories_management.php");
+					}
+				}
 			}
-		}else{
+		}
+		else {
 			$_SESSION['message_error'] = "Error - No se puede eliminar. Ingrese a 'categories_management.php'";
 			header("Location: categories_management.php");
 		}
-	}else {
+	}
+	else {
 		$_SESSION['message-error'] = "Acceso denegado.";
 		header("Location: products.php");
 	}
